services/finance-service/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9
django-environ>=0.11.2

# Async Support
asgiref>=3.7.2
uvicorn>=0.24.0
gunicorn>=21.2.0

# Celery & Redis
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1

# Payment Gateways
stripe>=7.0.0

# PDF Generation
weasyprint>=60.1
reportlab>=4.0.7

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# API Documentation
drf-spectacular>=0.26.5

# Monitoring & Logging
sentry-sdk>=1.34.0
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
flake8>=6.1.0
black>=23.11.0
isort>=5.12.0
